Arrange implicit factory test files
